web_chat/requirements.txt
# Web Chat - Dependencies
# Run from project root: pip install -r web_chat/requirements.txt

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.0
python-dotenv>=1.0.0
